feature: creating widgets from json
almost everything works, missing border for BorderWidget
